fix: moved SessionPayload to a separate file
