+++ src/app/lib/dtos/session.schema.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const SessionPayloadSchema = z.object({
+  id: z.bigint(),
+  role: z.string().optional(),
+  expiresAt: z.number(),
+});
+
+export type SessionPayload = z.infer<typeof SessionPayloadSchema>;
