+++ src/app/lib/services/session.ts
@@ -1,6 +1,6 @@
 import "server-only";
 import { SignJWT, jwtVerify } from "jose";
-import { SessionPayload } from "@/dtos/citizen.schema";
+import { SessionPayload } from "@/dtos/session.schema";
 import { cookies } from "next/headers";
 
 const secretKey = process.env.SESSION_SECRET;
